Moved prim implemetnations into rev due to the use of autodiff to compute the Jacobians
