+++ stan/math/rev/arr/functor/coupled_ode_system_cvode.hpp
@@ -1,5 +1,5 @@
-#ifndef STAN_MATH_PRIM_ARR_FUNCTOR_COUPLED_ODE_SYSTEM_CVODE_HPP
-#define STAN_MATH_PRIM_ARR_FUNCTOR_COUPLED_ODE_SYSTEM_CVODE_HPP
+#ifndef STAN_MATH_REV_ARR_FUNCTOR_COUPLED_ODE_SYSTEM_CVODE_HPP
+#define STAN_MATH_REV_ARR_FUNCTOR_COUPLED_ODE_SYSTEM_CVODE_HPP
 
 #include <stan/math/prim/arr/functor/coupled_ode_system.hpp>
