+++ test/unit/math/rev/arr/functor/coupled_ode_system_cvode_rev_test.cpp
@@ -1,7 +1,7 @@
 #include <gtest/gtest.h>
 #include <stan/math/rev/core.hpp>
 #include <stan/math/rev/arr/functor/coupled_ode_system.hpp>
-#include <stan/math/prim/arr/functor/coupled_ode_system_cvode.hpp>
+#include <stan/math/rev/arr/functor/coupled_ode_system_cvode.hpp>
 #include <test/unit/util.hpp>
 #include <test/unit/math/prim/arr/functor/harmonic_oscillator.hpp>
 #include <test/unit/math/prim/arr/functor/mock_ode_functor.hpp>
